+++ template/src/index.js
@@ -1,127 +1,10 @@
 // Copyright (c) 2018 Dirk Schumacher, Noam Ross, Rich FitzJohn
 // Copyright (c) 2025 Jinhwan Kim
 const { app, session, BrowserWindow } = require("electron");
-const { spawn, execSync } = require("child_process");
 const path = require("path");
 const os = require("os");
-
-// Error Handling System
-class ErrorHandler {
-  static logError(context, error, metadata = {}) {
-    const errorInfo = {
-      context,
-      message: error.message,
-      stack: error.stack,
-      timestamp: new Date().toISOString(),
-      ...metadata,
-    };
-
-    console.error(`[${context}] Error:`, errorInfo);
-
-    // TODO: Integrate with error reporting service in the future
-    // this.reportToService(errorInfo);
-  }
-
-  static async handleAsyncError(context, asyncFn, fallbackValue = null) {
-    try {
-      return await asyncFn();
-    } catch (error) {
-      this.logError(context, error);
-      return fallbackValue;
-    }
-  }
-
-  static handleSyncError(context, syncFn, fallbackValue = null) {
-    try {
-      return syncFn();
-    } catch (error) {
-      this.logError(context, error);
-      return fallbackValue;
-    }
-  }
-}
-
-// Process Management with better error handling
-class ProcessManager {
-  static async terminateRProcesses() {
-    if (os.platform() !== "win32") {
-      return;
-    }
-
-    return ErrorHandler.handleAsyncError(
-      "ProcessManager.terminateRProcesses",
-      async () => {
-        execSync("taskkill /IM Rterm.exe /F", { stdio: "ignore" });
-        console.log("Rterm.exe processes have been terminated.");
-      }
-    );
-  }
-
-  static async startShinyProcess(appState) {
-    return new Promise((resolve, reject) => {
-      const rShinyProcess = spawn(
-        appState.paths.rscript,
-        ["--vanilla", "-f", path.join(app.getAppPath(), "start-shiny.R")],
-        {
-          env: {
-            WITHIN_ELECTRON: "1",
-            RE_SHINY_PATH: appState.paths.shinyAppPath,
-            R_LIB_PATHS: appState.paths.libPath,
-            R_HOME_DIR: appState.paths.rpath,
-            RHOME: appState.paths.rpath,
-            RE_SHINY_PORT: appState.config.serverPort,
-            R_LIBS: appState.paths.libPath,
-            R_LIBS_USER: appState.paths.libPath,
-            R_LIBS_SITE: appState.paths.libPath,
-          },
-          stdio: "ignore",
-        }
-      );
-
-      rShinyProcess.on("error", (err) => {
-        ErrorHandler.logError("ProcessManager.startShinyProcess", err, {
-          pid: rShinyProcess.pid,
-          command: appState.paths.rscript,
-        });
-        reject(err);
-      });
-
-      rShinyProcess.on("exit", (code, signal) => {
-        const exitInfo = { code, signal, pid: rShinyProcess.pid };
-
-        if (code === 0) {
-          console.log("Shiny process exited normally");
-          resolve(rShinyProcess);
-        } else {
-          const error = new Error(`Shiny process exited with code ${code}`);
-          ErrorHandler.logError(
-            "ProcessManager.startShinyProcess",
-            error,
-            exitInfo
-          );
-          reject(error);
-        }
-      });
-
-      // Resolve after successful start
-      setTimeout(() => resolve(rShinyProcess), 100);
-    });
-  }
-
-  static async killProcess(process, processName = "process") {
-    if (!process) {
-      return;
-    }
-
-    return ErrorHandler.handleAsyncError(
-      `ProcessManager.killProcess.${processName}`,
-      async () => {
-        process.kill();
-        console.log(`${processName} killed successfully`);
-      }
-    );
-  }
-}
+const ErrorHandler = require("./error-handler");
+const ProcessManager = require("./process-manager"); // <-- import ProcessManager
 
 // Application State Management with better error handling
 class AppState {
